wish_list-main/book_wishlist-main/src/requirements.txt
analytics-python==1.2.9
aniso8601==8.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
flask-segmentio==0.1.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
json-logging==1.2.10
MarkupSafe==1.1.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sentry-sdk[flask]==0.20.2
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
Werkzeug==1.0.1
gunicorn
psycopg2-binary
